requirements.txt
apify-client
google-generativeai
python-dotenv
requests
pyautogen
